# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Cantina da Sandra",
    "description": "API para o sistema de pedidos da cantina da Sandra"
  },
  "host": "localhost:4000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "scheme": "bearer",
      "bearerFormat": "JWT"
    }
  },
  "paths": {
...
    "/produtos/{id}": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "description": "Endpoint para obter um produto.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Id do produto"
          }
        ],
        "responses": {
          "200": {
            "description": "Produto",
            "schema": {
              "type": "object",
              "properties": {
                "Id": {
                  "type": "string",
                  "example": "number"
                },
                "Nome": {
                  "type": "string",
                  "example": "string"
                },
                "Preco": {
                  "type": "string",
                  "example": "number"
                },
                "Descricao": {
                  "type": "string",
                  "example": "string"
                },
                "URLImagem": {
                  "type": "string",
                  "example": "string"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Produto não encontrado"
          }
        }
      }
    },
    "/produtos/adicionar": {
      "post": {
        "tags": [
          "Produtos"
        ],
        "description": "Endpoint para adicionar um produto.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Token de autenticação",
            "type": "string"
          },
          {
            "name": "produto",
            "in": "body",
            "description": "Produto a ser adicionado",
            "schema": {
              "type": "object",
              "properties": {
                "Nome": {
                  "type": "string",
                  "example": "string"
                },
                "Preco": {
                  "type": "string",
                  "example": "number"
                },
                "Descricao": {
                  "type": "string",
                  "example": "string"
                },
                "URLImagem": {
                  "type": "string",
                  "example": "string"
                },
                "Categoria": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto adicionado"
          },
          "400": {
            "description": "Erro ao adicionar produto"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/produtos/atualizar/{id}": {
      "put": {
        "tags": [
          "Produtos"
        ],
        "description": "Endpoint para atualizar um produto.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Id do produto"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Token de autenticação",
            "type": "string"
          },
          {
            "name": "produto",
            "in": "body",
            "description": "Produto a ser atualizado",
            "schema": {
              "type": "object",
              "properties": {
                "Nome": {
                  "type": "string",
                  "example": "string"
                },
                "Preco": {
                  "type": "string",
                  "example": "number"
                },
                "Descricao": {
                  "type": "string",
                  "example": "string"
                },
                "URLImagem": {
                  "type": "string",
                  "example": "string"
                },
                "Categoria": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto atualizado"
          },
          "400": {
            "description": "Erro ao atualizar produto"
          },
          "404": {
            "description": "Produto não encontrado"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/produtos/deletar/{id}": {
      "delete": {
        "tags": [
          "Produtos"
        ],
        "description": "Endpoint para deletar um produto.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Id do produto"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Token de autenticação",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Produto deletado"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Erro ao deletar produto"
          }
        }
      }
    }
  },
...
}